server/requirements.txt
flask
flask-cors
pdfplumber
google-generativeai
python-dotenv
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-bcrypt
flask-limiter
redis